+++ src/onegov.agency/HISTORY.rst
@@ -1,5 +1,7 @@
 Changelog
 ---------
+1.5.1 (2019-01-07)
+~~~~~~~~~~~~~~~~~~~~
 
 - Shows additional information for ambiguous people when editing memberships.
   [msom]